+++ src/main/java/com/brand/adabranium/Adabranium.java
@@ -1,9 +1,9 @@
 package com.brand.adabranium;
 
-import com.brand.adabranium.content.AdabraniumGroup;
-import com.brand.adabranium.content.ModBlocks;
-import com.brand.adabranium.content.ModItems;
-import com.brand.adabranium.content.stuff.potions.PotionsRecipes;
+import com.brand.adabranium.itemgroups.AdabraniumGroup;
+import com.brand.adabranium.registry.content.ModBlocks;
+import com.brand.adabranium.registry.content.ModItems;
+import com.brand.adabranium.registry.stuff.potions.PotionsRecipes;
 import com.brand.adabranium.worldgen.AdabraniumWorldgenFeatures;
 import net.fabricmc.api.ModInitializer;
 import net.minecraft.util.Identifier;
@@ -30,7 +30,7 @@ public class Adabranium implements ModInitializer {
     }
 
     public static Identifier id(String name) {
-        return new Identifier(MOD_ID, name);
+        return Identifier.of(MOD_ID, name);
     }
 }