+++ src/main/java/com/brand/adabranium/AdabraniumClient.java
@@ -1,14 +1,14 @@
 package com.brand.adabranium;
 
-import com.brand.adabranium.content.ModBlocks;
+import com.brand.adabranium.registry.content.ModBlocks;
 import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
-import net.minecraft.client.render.RenderLayer;
+import net.fabricmc.fabric.api.client.rendering.v1.BlockRenderLayerMap;
+import net.minecraft.client.render.BlockRenderLayer;
 
 public class AdabraniumClient implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
-        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.HEART_SHAPED_PLANT, RenderLayer.getCutoutMipped());
+        BlockRenderLayerMap.putBlocks(BlockRenderLayer.CUTOUT_MIPPED, ModBlocks.HEART_SHAPED_PLANT);
     }
 }