+++ src/main/java/com/brand/adabranium/blocks/HeartShapedPlantBlock.java
@@ -1,13 +1,9 @@
 package com.brand.adabranium.blocks;
 
-import com.brand.adabranium.content.ModBlocks;
-import com.brand.adabranium.content.ModItems;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.PlantBlock;
-import net.minecraft.block.ShapeContext;
+import com.brand.adabranium.registry.content.ModBlocks;
+import com.brand.adabranium.registry.content.ModItems;
+import com.mojang.serialization.MapCodec;
+import net.minecraft.block.*;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.world.ServerWorld;
@@ -17,78 +13,90 @@ import net.minecraft.state.StateManager;
 import net.minecraft.state.property.IntProperty;
 import net.minecraft.state.property.Properties;
 import net.minecraft.util.ActionResult;
-import net.minecraft.util.Hand;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.random.Random;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
-
-import java.util.Random;
+import net.minecraft.world.WorldView;
+import net.minecraft.world.event.GameEvent;
 
 public class HeartShapedPlantBlock extends PlantBlock {
+    public static final MapCodec<SweetBerryBushBlock> CODEC = createCodec(SweetBerryBushBlock::new);
+    public static final int MAX_AGE = 2;
     public static final IntProperty AGE;
-    protected static final VoxelShape[] AGE_TO_SHAPE;
+    public static final VoxelShape SMALL_SHAPE;
+    public static final VoxelShape LARGE_SHAPE;
 
-    public HeartShapedPlantBlock(Settings settings) {
+    public MapCodec<SweetBerryBushBlock> getCodec() {
+        return CODEC;
+    }
+
+    public HeartShapedPlantBlock(AbstractBlock.Settings settings) {
         super(settings);
         this.setDefaultState(this.stateManager.getDefaultState().with(AGE, 0));
     }
 
-    @Environment(EnvType.CLIENT)
-    public ItemStack getPickStack(BlockView world, BlockPos pos, BlockState state) {
+    protected ItemStack getPickStack(WorldView world, BlockPos pos, BlockState state, boolean includeData) {
         return new ItemStack(ModBlocks.HEART_SHAPED_PLANT);
     }
 
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        Vec3d vec3d = state.getModelOffset(world, pos);
-        return AGE_TO_SHAPE[state.get(this.getAgeProperty())].offset(vec3d.x, vec3d.y, vec3d.z);
+        if (state.get(AGE) == 0) {
+            return SMALL_SHAPE;
+        } else {
+            return state.get(AGE) < 2 ? LARGE_SHAPE : super.getOutlineShape(state, world, pos, context);
+        }
     }
 
-    public OffsetType getOffsetType() {
-        return OffsetType.XZ;
-    }
-
-    public IntProperty getAgeProperty() {
-        return AGE;
+    public boolean hasRandomTicks(BlockState state) {
+        return state.get(AGE) < 2;
     }
 
     public void randomTick(BlockState state, ServerWorld world, BlockPos pos, Random random) {
         int i = state.get(AGE);
         if (i < 2 && random.nextInt(5) == 0 && world.getBaseLightLevel(pos.up(), 0) >= 10) {
-            world.setBlockState(pos, state.with(AGE, i + 1), 2);
+            BlockState blockState = state.with(AGE, i + 1);
+            world.setBlockState(pos, blockState, 2);
+            world.emitGameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Emitter.of(blockState));
         }
 
     }
 
-    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, BlockHitResult hit) {
         int i = state.get(AGE);
+        boolean bl = i == 2;
         if (i > 1) {
+            int j = 1 + world.random.nextInt(2);
             dropStack(world, pos, new ItemStack(ModItems.HEART_SHAPED_HERB, 1));
             world.playSound(null, pos, SoundEvents.BLOCK_SWEET_BERRY_BUSH_PICK_BERRIES, SoundCategory.BLOCKS, 1.0F, 0.8F + world.random.nextFloat() * 0.4F);
-            world.setBlockState(pos, state.with(AGE, 0), 2);
+            BlockState blockState = state.with(AGE, 1);
+            world.setBlockState(pos, blockState, 2);
+            world.emitGameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Emitter.of(player, blockState));
             return ActionResult.SUCCESS;
         } else {
-            return super.onUse(state, world, pos, player, hand, hit);
+            return super.onUse(state, world, pos, player, hit);
         }
     }
 
-    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
+    public void appendProperties(StateManager.Builder<Block, BlockState> builder) {
         builder.add(AGE);
     }
+
 
     public boolean canGrow(World world, Random random, BlockPos pos, BlockState state) {
         return true;
     }
 
     public void grow(ServerWorld world, Random random, BlockPos pos, BlockState state) {
-        int i = Math.min(2, state.get(AGE) + 1);
-        world.setBlockState(pos, state.with(AGE, i), 2);
+        int i = Math.min(3, (Integer)state.get(AGE) + 1);
+        world.setBlockState(pos, (BlockState)state.with(AGE, i), 2);
     }
 
     static {
         AGE = Properties.AGE_2;
-        AGE_TO_SHAPE = new VoxelShape[]{Block.createCuboidShape(6.0D, 0.0D, 6.0D, 10.0D, 10.0D, 10.0D), Block.createCuboidShape(4.0D, 0.0D, 4.0D, 12.0D, 12.0D, 12.0D), Block.createCuboidShape(2.0D, 0.0D, 2.0D, 14.0D, 15.0D, 14.0D)};
+        SMALL_SHAPE = Block.createCuboidShape(3.0, 0.0, 3.0, 13.0, 8.0, 13.0);
+        LARGE_SHAPE = Block.createCuboidShape(1.0, 0.0, 1.0, 15.0, 16.0, 15.0);
     }
 }