+++ changelog.md
@@ -1,2 +1,7 @@
-
-* Fix Crash at startup
+* Updated to 1.21.7
+* Vibranium & Adamantium Ores now require Netherite Pickaxe tier to be obtained
+* Vibranium Ingot is now crafted using 4 Vibranium & 4 Obsidian
+* Adamantium Ingot is now crafted using 4 Adamantium & 4 Vibranium
+* Adamantium Stuff is now obtained by upgrading vibranium stuff
+  * Removed now unused Adamantium
+* Reduced Vibranium & Adamantium stuff durability